Fixed an error in how the fully connected layer was finding and instantiating the helper functions in the cpp file.
Templated the weights, optimizers, objective functions, and objective
function terms with TensorDataType.

Note that this temporarily spilled TensorDataType to the model.  The
model should be updated to only hold layers, and layers hold the
weights, optimizers, and objective functions.  This will prevent the
data type from polluting the model class, but is for a separate PR.